chore: rename rule to `ad-hoc-sub-process`
