+++ test/config/configs.spec.js
@@ -401,7 +401,7 @@ describe('configs', function() {
 
 
   it('camunda-cloud-8-7', expectRules(configs, 'camunda-cloud-8-7', {
-    'ad-hoc-sub-process-activity': [ 'error', { version: '8.7' } ],
+    'ad-hoc-sub-process': [ 'error', { version: '8.7' } ],
     'called-element': [ 'error', { version: '8.7' } ],
     'connector-properties': [ 'warn', { version: '8.7' } ],
     'duplicate-execution-listeners': [ 'error', { version: '8.7' } ],
@@ -439,7 +439,7 @@ describe('configs', function() {
 
 
   it('camunda-cloud-8-8', expectRules(configs, 'camunda-cloud-8-8', {
-    'ad-hoc-sub-process-activity': [ 'error', { version: '8.8' } ],
+    'ad-hoc-sub-process': [ 'error', { version: '8.8' } ],
     'called-element': [ 'error', { version: '8.8' } ],
     'connector-properties': [ 'warn', { version: '8.8' } ],
     'duplicate-execution-listeners': [ 'error', { version: '8.8' } ],
@@ -501,7 +501,7 @@ describe('configs', function() {
 
 
   it('all', expectRules(configs, 'all', {
-    'ad-hoc-sub-process-activity': 'error',
+    'ad-hoc-sub-process': 'error',
     'element-type': 'error',
     'called-element': 'error',
     'collapsed-subprocess': 'error',
